+++ src/main/java/smile/identity/core/adapters/InstantAdapter.java
@@ -3,6 +3,7 @@ package smile.identity.core.adapters;
 import com.squareup.moshi.FromJson;
 import com.squareup.moshi.ToJson;
 
+import java.time.DateTimeException;
 import java.time.Instant;
 
 public class InstantAdapter {
@@ -11,7 +12,13 @@ public class InstantAdapter {
         return instant.toString();
     }
 
-    @FromJson Instant fromJson(String string){
-        return Instant.parse(string);
+    @FromJson Instant fromJson(String string) {
+        Instant timestamp;
+        try {
+            timestamp = Instant.ofEpochMilli(Long.parseLong(string));
+        } catch (DateTimeException | NumberFormatException ex) {
+            timestamp = Instant.parse(string);
+        }
+        return timestamp;
     }
 }