+++ changelog.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [2.0.4] - 2023-3-7
+* JobResponse now returns result as an Object
+* New methods to get result as String or JobResponse
+* Updated parsing of Instant
+
 ## [2.0.3 ] - 2023-2-13
 * Added logging for error responses from server
 * Added method to validate signature from string timestamp